+++ CHANGES.md
@@ -12,6 +12,7 @@ v1.2.0 - Pending
 - Fixed configure script not using warning/preprocessor options in build
   (Issue #60)
 - Fixed printer renaming algorithm to not truncate the name (Issue #60)
+- Fixed missing "print-color-mode=bi-level" for bar code printing (Issue #76)
 - Fixed label mode support in EPL and ZPL drivers (Issue #79)
 - Fixed driver names for EPL printers (Issue #52)